+++ src/main/java/com/tabletmc/echo_summon/ClientInitializer.java
@@ -1,5 +1,7 @@
 package com.tabletmc.echo_summon;
 import com.tabletmc.echo_summon.client.tooltip.SaddleSummonToolTooltipClient;
+import com.tabletmc.echo_summon.client.tooltip.HarnessSummonToolTooltipClient;
+import com.tabletmc.echo_summon.client.tooltip.EquipmentTooltipsClient;
 import com.tabletmc.echo_summon.keybinds.KeybindTickEvents;
 import com.tabletmc.echo_summon.keybinds.RegisterKeybinds;
 import com.tabletmc.echo_summon.net.ClientNetworking;
@@ -12,6 +14,8 @@ public class ClientInitializer implements net.fabricmc.api.ClientModInitializer 
         KeybindTickEvents.init();
         RegisterKeybinds.ALL.forEach(KeyBindingHelper::registerKeyBinding);
         SaddleSummonToolTooltipClient.register();
+        HarnessSummonToolTooltipClient.register();
+        EquipmentTooltipsClient.register();
         // TODO: Re-implement mount summon tool model variants using the 1.21+ item asset system instead of runtime predicates.
     }
 }